Removed excees comments
