+++ libruse-read/src/lex/error.rs
@@ -1,31 +1,19 @@
-//! Lexing error and result types.
-
 use lex::token::{Token, Location};
-
 use std::error;
 use std::fmt;
 use std::result;
 
-/// The result of lexing. Either a vector of tokens if the lexing is successful
-/// or an Error if it's not.
 pub type Result = result::Result<Vec<Token>, Error>;
 
-/// A lexing error, indicating some problem with the contents of the input
-/// stream.
 #[derive(Debug, PartialEq)]
 pub enum Error {
-    /// Character and location
     InvalidCharacter(char, Location),
-    /// The almost-number.
     MalformedNumber(String, Location),
-    /// An almost-literal
     InvalidLiteral(String, Location),
-    /// An almost-escape sequence
     InvalidEscapeSequence(String, Location),
 }
 
 impl error::Error for Error {
-    /// Get a simple text description of what each error means.
     fn description(&self) -> &str {
         match *self {
             Error::InvalidCharacter(..) => "invalid character",
@@ -37,7 +25,6 @@ impl error::Error for Error {
 }
 
 impl fmt::Display for Error {
-    /// Print detailed error information.
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match *self {
             Error::InvalidCharacter(character, location) => {
