+++ libruse-read/src/lex/token.rs
@@ -1,38 +1,24 @@
-//! A single lexical item.
-
 use lex::error::Error;
-
 use std::cell::Cell;
 use std::iter::Peekable;
 use std::str::Chars;
 use std::iter::Iterator;
 use std::fmt;
 
-/// The different types of Token. Some token kinds require additional data,
-/// and will carry it here.
 #[derive(Debug, PartialEq, Clone)]
 pub enum TokenKind {
-    /// A single open delimiter.
     OpenDelim(Delim),
-    /// A single close delimiter.
     CloseDelim(Delim),
-    /// An identifier.
     Ident(String),
-    /// An integer literal.
     Integer(i64),
-    /// A floating point literal.
     Float(f64),
-    /// A string
     Str(String),
-    /// A boolean
     Bool(bool),
     // TODO: Implement block comments
-    /// A block comment
     BlockComment(String),
 }
 
 impl TokenKind {
-    /// Checks whether the token kind is an open delim.
     pub fn is_open_delim(&self) -> bool {
         match *self {
             TokenKind::OpenDelim(..) => true,
@@ -41,18 +27,13 @@ impl TokenKind {
     }
 }
 
-/// Indicates type of delimiter.
 #[derive(PartialEq, Eq, Debug, Clone)]
 pub enum Delim {
-    /// Parens: ( and )
     Paren,
-    /// Brackets: [ and ]
     Bracket,
-    /// Braces: { and }
     Brace,
 }
 
-/// The location of a token in an input stream.
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub struct Location {
     column: usize,
@@ -74,22 +55,15 @@ impl fmt::Display for Location {
     }
 }
 
-/// Tokens, generated by the lexer. Every token has a location in the input
-/// stream, a span indicating how large they are, and a type that potentially
-/// has some associated data.
 #[derive(Debug, PartialEq)]
 pub struct Token {
-    /// The kind of token, potentially with some additional data.
     pub kind: TokenKind,
-    /// The start location of the token in the input stream.
     pub start_location: Location,
-    /// The end location of the token in the input stream.
     pub end_location: Location,
 }
 
 macro_rules! delim_token {
     ( $fn_name:ident, $kind:expr ) => {
-        /// Create a delimiter token.
         pub fn $fn_name(start_location: Location, end_location: Location) -> Token {
             Token {
                 kind: $kind,
@@ -102,7 +76,6 @@ macro_rules! delim_token {
 
 macro_rules! stringy_token {
     ( $fn_name:ident, $variant_name:ident ) => {
-        /// Create a stringy token.
         pub fn $fn_name<S: Into<String>>(s: S, start_location: Location) -> Token {
             let s: String = s.into();
 
@@ -122,7 +95,6 @@ macro_rules! stringy_token {
 
 macro_rules! literal_token {
     ( $fn_name:ident, $variant_name:ident, $value_type:ty ) => {
-        /// Create a literal token.
         pub fn $fn_name(value: $value_type,
                         start_location: Location,
                         end_location: Location) -> Token {
@@ -151,22 +123,17 @@ impl Token {
     literal_token!(boolean, Bool, bool);
 }
 
-/// An iterator over the tokens in an input string.
 pub struct TokenIterator<'a> {
-    /// A peekable iterator over the characters in the original string.
     char_iter: Peekable<Chars<'a>>,
-    /// The location of the iterator in the input stream.
     location: Cell<Location>,
 }
 
-/// Indicates whether to move the internal character iterator.
 enum IterateInternally {
     Yes,
     No,
 }
 
 impl<'a> TokenIterator<'a> {
-    /// Create a new TokenIterator to iterate over the given string.
     pub fn new(s: &str) -> TokenIterator {
         TokenIterator {
             char_iter: s.chars().peekable(),
@@ -174,7 +141,6 @@ impl<'a> TokenIterator<'a> {
         }
     }
 
-    /// Increment the iterator's internal location field.
     fn next_location(&mut self, go_to_next: IterateInternally) {
         if self.char_iter.peek().filter(|v| **v == '\n').is_some() {
             let new_location = Location {
@@ -190,14 +156,12 @@ impl<'a> TokenIterator<'a> {
         }
     }
 
-    /// Get the current location of the parser in the input text.
     fn location(&self) -> Location {
         self.location.get()
     }
 }
 
 impl<'a> Iterator for TokenIterator<'a> {
-    /// Returns either a Token or a lexing error.
     type Item = Result<Token, Error>;
 
     /// Returns one of three things:
@@ -238,9 +202,7 @@ impl<'a> Iterator for TokenIterator<'a> {
     }
 }
 
-/// Extend a stringy type with the ability to generate tokens.
 pub trait StrTokenIterator: AsRef<str> {
-    /// Convenience method to get a token iterator for a string.
     fn tokens(&self) -> TokenIterator {
         TokenIterator::new(self.as_ref())
     }
@@ -270,12 +232,6 @@ lex_delim!(lex_closed_bracket, close_bracket, 1);
 lex_delim!(lex_open_brace, open_brace, 1);
 lex_delim!(lex_closed_brace, close_brace, 1);
 
-/// Parse a number, either floating point or integer.
-///
-/// Attempts to parse a number. For now, it parses only integers and
-/// floating point numbers. Eventually we'll want to cover the entire
-/// numeric tower that Scheme requires, which will demand substantially
-/// more work than is being done now.
 fn lex_number(iter: &mut TokenIterator, character: char) -> Result<Token, Error> {
     let mut result = vec![character];
     let start = iter.location();
@@ -302,14 +258,6 @@ fn lex_number(iter: &mut TokenIterator, character: char) -> Result<Token, Error>
     }
 }
 
-/// Parse an ident.
-///
-/// This parses an ident, starting with the given character, which is
-/// passed in for convenience. Note that at the moment the character
-/// matching rules are kept in sync manually between this function and the
-/// original dispatcher, and that they are not exactly the same. There are
-/// certain characters which are acceptible within an identifier that are
-/// not acceptable at the start of one.
 fn lex_ident(iter: &mut TokenIterator, character: char) -> Result<Token, Error> {
     let mut result = vec![character];
     let start = iter.location();
@@ -332,7 +280,6 @@ fn lex_ident(iter: &mut TokenIterator, character: char) -> Result<Token, Error> 
     Ok(Token::ident(out, start))
 }
 
-/// Lex a boolean value.
 fn lex_boolean(iter: &mut TokenIterator, character: char) -> Result<Token, Error> {
     let mut result = vec![character];
     let start = iter.location();
@@ -361,7 +308,6 @@ fn lex_boolean(iter: &mut TokenIterator, character: char) -> Result<Token, Error
     }
 }
 
-/// Lex a string literal.
 fn lex_string(iter: &mut TokenIterator, character: char) -> Result<Token, Error> {
     let mut result = Vec::new();
     let mut escape = false;
